backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
langchain==0.3.13
langgraph==0.2.60
langchain-google-genai==2.0.7
langchain-community==0.3.13
sqlalchemy==2.0.36
pydantic==2.10.3
httpx==0.27.2
python-dotenv==1.0.1
PyJWT==2.9.0
python-multipart==0.0.17
